Move the style generator into a separate file
