+++ src/components/ui/radio-group.svelte
@@ -1,4 +1,6 @@
 <script>
+  import getColorPickerStyles from './utils/color-picker-styles.js';
+
   export let classname = '';
   export let wrapperclass = 'radio';
   export let radioclass = 'radio';
@@ -24,30 +26,6 @@
   if (labelPosition !== 'right' && labelPosition !== 'left') {
     throw new Error('Label position must be either left or right.');
   }
-
-  function radioColor(hexColor) {
-    if (!isColor) return '';
-
-    if (hexColor.length !== 7) {
-      throw new Error('Values of colored radio groups must be #XXXXXX hex colors.');
-    }
-
-    let r = parseInt(hexColor.slice(1, 3), 16);
-    let g = parseInt(hexColor.slice(3, 5), 16);
-    let b = parseInt(hexColor.slice(5, 7), 16);
-
-    const whiteThreshold = 245;
-    const gray = 153;
-    if (r > whiteThreshold && g > whiteThreshold && b > whiteThreshold) {
-      r = gray;
-      g = gray;
-      b = gray;
-    }
-
-    return `background-color: rgb(${r}, ${g}, ${b});
-            border-color: rgb(${r}, ${g}, ${b});
-            --r: ${r}; --g: ${g}; --b: ${b};`;
-  }
 </script>
 
 <div class:with-labels={!isColor} class={classname} role="group">
@@ -69,7 +47,9 @@
               name={name}
               class={inputclass}
           >
-          <div style="{radioColor(loopValue)}" class={iconclass}></div>
+          <div
+            style="{isColor ? getColorPickerStyles(loopValue) : ''}"
+            class="icon {iconclass}"></div>
         </div>
       {:else}
         <input
@@ -79,7 +59,9 @@
             name={name}
             class={inputclass}
         >
-        <div style="{radioColor(loopValue)}" class={iconclass}></div>
+        <div
+          style="{isColor ? getColorPickerStyles(loopValue) : ''}"
+          class="icon {iconclass}"></div>
       {/if}
       {#if !isColor && labelPosition === 'right'}
         {#if labelclass !== null}
